% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{add_two_vectors}
\alias{add_two_vectors}
\title{Add Two Vectors}
\usage{
add_two_vectors(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A numeric vector containing the element-wise sum of x and y
}
\description{
This function takes two input vectors and returns their element-wise sum. The two input vectors
must have the same length.
}
