% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_other_function.R
\name{another_fun}
\alias{another_fun}
\title{Another Function}
\usage{
another_fun(x)
}
\arguments{
\item{x}{numeric input}
}
\value{
numeric output
}
\description{
adds 2 to input
}
\examples{
another_fun(1)
}
