% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunction.R
\name{dummyFunction}
\alias{dummyFunction}
\title{Prints Hello World!}
\usage{
dummyFunction()
}
\value{
The sentence "Hello, world!
}
\description{
A light package to use as a dummy and light example
}
\examples{
dummyFunction()
}
