\name{dummypackage-package}
\alias{dummypackage-package}
\alias{dummypackage}
\docType{package}
\title{
\packageTitle{dummypackage}
}
\description{
\packageDescription{dummypackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dummypackage}
\packageIndices{dummypackage}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{dummypackage}

Maintainer: \packageMaintainer{dummypackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
