% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{generate_random_matrix}
\alias{generate_random_matrix}
\title{Generate a Random Matrix}
\usage{
generate_random_matrix(nrows, ncols)
}
\arguments{
\item{nrows}{The number of rows in the matrix, which should be a positive integer}

\item{ncols}{The number of columns in the matrix, which should be a positive integer}
}
\value{
A matrix of dimensions nrows x ncols, with each element drawn from a standard normal distribution.
}
\description{
This function generates a random matrix of the specified dimensions, with each element
being drawn from a standard normal distribution.
}
