% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{generate_random_vector}
\alias{generate_random_vector}
\title{Generate a Random Vector}
\usage{
generate_random_vector(length)
}
\arguments{
\item{length}{The length of the desired vector, which should be a positive integer.}
}
\value{
A vector of length 'length', containing random values between 0 and 1.
}
\description{
This function generates a random vector of a specified length using the runif() function.
}
