% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{matrix_multiplication}
\alias{matrix_multiplication}
\title{Perform Matrix Multiplication}
\usage{
matrix_multiplication(a, b)
}
\arguments{
\item{a}{A matrix}

\item{b}{A matrix}
}
\value{
The result of matrix multiplication of a and b
}
\description{
This function performs matrix multiplication of two matrices. The two input matrices
must have compatible dimensions for multiplication.
}
