% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummypackage-package.R
\name{ols_dbl}
\alias{ols_dbl}
\title{OLS Estimator as a Double}
\usage{
ols_dbl(y, x)
}
\arguments{
\item{y}{Column vector of dependent variable}

\item{x}{Matrix of independent variables}
}
\value{
Double vector of OLS estimates
}
\description{
OLS Estimator as a Double
}
\examples{
y <- mtcars_mat$y
X <- mtcars_mat$x[, c("wt", "hp")]
X <- cbind(1, X)
ols_dbl(y, X)
}
