% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummypackage-package.R
\name{ols_mat}
\alias{ols_mat}
\title{OLS Estimator as a Matrix}
\usage{
ols_mat(y, x)
}
\arguments{
\item{y}{Column vector of dependent variable}

\item{x}{Matrix of independent variables}
}
\value{
Matrix of OLS estimates
}
\description{
OLS Estimator as a Matrix
}
\examples{
y <- mtcars_mat$y
X <- mtcars_mat$x[, c("wt", "hp")]
X <- cbind(1, X)
ols_mat(y, X)
}
