% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{paste_matrices}
\alias{paste_matrices}
\title{Concatenate Matrices or Vectors}
\usage{
paste_matrices(x, y)
}
\arguments{
\item{x}{A vector or matrix}

\item{y}{A vector or matrix}
}
\value{
A matrix, created by horizontally concatenating x and y
}
\description{
This function concatenates two matrices or vectors horizontally. If both inputs are vectors,
they are concatenated vertically to create a 2-column matrix. If both inputs are matrices,
they are concatenated horizontally using cbind(). If the inputs are not both vectors or both matrices,
the function will stop execution and return an error message.
}
