% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{dummy_fun}
\alias{dummy_fun}
\title{Dummy Function}
\usage{
dummy_fun(x)
}
\arguments{
\item{x}{numeric input}
}
\value{
numeric output
}
\description{
adds 1 to the input
}
\examples{
my_function(1)
my_function(10)
}
