% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummypackage-package.R
\docType{package}
\name{dummypackage-package}
\alias{dummypackage}
\alias{dummypackage-package}
\title{dummypackage: Just a Dummy Package to Show Wide-Range Testing}
\description{
This package provides an overkill by adding R and C++ code to show how to test a package with a wide range of tests on different operating systems with different compilers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/dummypackage}
  \item Report bugs at \url{https://github.com/pachadotdev/dummypackage/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
